admin: split announcement tasks
