+++ handlers/admin/delete_announcement_task.go
@@ -1,0 +1,63 @@
+package admin
+
+import (
+	"fmt"
+	"net/http"
+	"strconv"
+
+	"github.com/arran4/goa4web/core/common"
+	"github.com/arran4/goa4web/core/consts"
+	"github.com/arran4/goa4web/handlers"
+	notif "github.com/arran4/goa4web/internal/notifications"
+	"github.com/arran4/goa4web/internal/tasks"
+)
+
+// DeleteAnnouncementTask removes an announcement.
+type DeleteAnnouncementTask struct{ tasks.TaskString }
+
+var deleteAnnouncementTask = &DeleteAnnouncementTask{TaskString: TaskDelete}
+
+var _ tasks.Task = (*DeleteAnnouncementTask)(nil)
+var _ tasks.AuditableTask = (*DeleteAnnouncementTask)(nil)
+
+// deleteAnnouncementTask also notifies admins of changes for transparency.
+var _ notif.AdminEmailTemplateProvider = (*DeleteAnnouncementTask)(nil)
+
+func (DeleteAnnouncementTask) Action(w http.ResponseWriter, r *http.Request) any {
+	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
+	if err := r.ParseForm(); err != nil {
+		return fmt.Errorf("parse form fail %w", handlers.ErrRedirectOnSamePageHandler(err))
+	}
+	for _, idStr := range r.Form["id"] {
+		id, _ := strconv.Atoi(idStr)
+		if err := queries.DeleteAnnouncement(r.Context(), int32(id)); err != nil {
+			return fmt.Errorf("delete announcement fail %w", handlers.ErrRedirectOnSamePageHandler(err))
+		}
+		if cd, ok := r.Context().Value(consts.KeyCoreData).(*common.CoreData); ok {
+			if evt := cd.Event(); evt != nil {
+				if evt.Data == nil {
+					evt.Data = map[string]any{}
+				}
+				evt.Data["AnnouncementID"] = id
+			}
+		}
+	}
+	return nil
+}
+
+func (DeleteAnnouncementTask) AdminEmailTemplate() *notif.EmailTemplates {
+	return notif.NewEmailTemplates("announcementEmail")
+}
+
+func (DeleteAnnouncementTask) AdminInternalNotificationTemplate() *string {
+	v := notif.NotificationTemplateFilenameGenerator("announcement")
+	return &v
+}
+
+// AuditRecord summarises an announcement deletion.
+func (DeleteAnnouncementTask) AuditRecord(data map[string]any) string {
+	if id, ok := data["AnnouncementID"].(int); ok {
+		return fmt.Sprintf("announcement %d deleted", id)
+	}
+	return "announcement deleted"
+}
